fixed bug in standard similarity
* pair classes moved to utils
* when computing standard similarity RCL coefficients were using
larger sizes of clusters (size after merging)
